配置文件

+++ config/autoload/http.go
@@ -2,5 +2,5 @@ package autoload
 
 type Http struct {
 	Host string `yaml:"Host"`
-	Port int    `yaml:"Port"`
+	Port uint16 `yaml:"Port"`
 }
